Rename the file facade01config.py
There is no reason to have them prefixed like this, and it make harder
to understand what is going on
