+++ augur/tasks/git/util/facade_worker/facade_worker/facade00mainprogram.py
@@ -29,7 +29,7 @@ from __future__ import annotations
 import traceback
 import sys, platform, imp, time, datetime, html.parser, subprocess, os, getopt, xlsxwriter, configparser, logging
 from multiprocessing import Process, Queue
-from .facade01config import FacadeSession as FacadeSession#increment_db, update_db, migrate_database_config, database_connection, get_setting, update_status, log_activity          
+from .config import FacadeSession as FacadeSession
 from .facade02utilitymethods import trim_commit, store_working_author, trim_author   
 from .facade03analyzecommit import analyze_commit
 from .facade04postanalysiscleanup import git_repo_cleanup
